Parser for qualified tags
Implements a function that parses a string containing a fully-qualified
tag name (ie. a sequence of tag names and '.' field separators or array
indexes) into their constituent parts.

Part of WTP-19: each constituent string returned by the parser will be used to
identify a node in the tag locktree.

Splitting this into its own patch because we may want this to live farther
up the stack towards the API, since presumably we'll need something like
this for the typechecking functionality.  If that's the case, the return type
perhaps shouldn't be a slice of strings but perhaps some other tagtype-aware
datatype.