Tag parser, v2 (#17)
Looks like james/tag_parser_multidim_arrays had a few failing
tests which I only noticed after merging to master.  My fault
for merging.  But, I took this opportunity to rethink how we
should be parsing the tags and refactored the one large function
into several smaller ones, one for each EBNF production.

I think this is cleaner and likely more maintainable and less
spaghetti-ish.

Some of the tests needed to be changed, but only in terms of what
particular error is produced to the user; the functionality should
still be what we want.